% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enigma.R
\docType{methods}
\name{encrypt}
\alias{encrypt}
\alias{encrypt,Enigma-method}
\title{Encrypt a string using an Enigma instance}
\usage{
encrypt(Enigma, string)

\S4method{encrypt}{Enigma}(Enigma, string)
}
\arguments{
\item{Enigma}{An object of class 'Enigma'}

\item{string}{Character string to encrypt.}
}
\description{
Encrypt a string using an Enigma instance
}
