% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enigma.R
\name{Enigma}
\alias{Enigma}
\title{Initialize an M3 Enigma machine}
\usage{
Enigma(rotors = c("III", "II", "I"), start_positions = c("A", "A",
  "A"), ring_settings = c("A", "A", "A"), reflector = "B",
  plugboard = c())
}
\arguments{
\item{rotors}{Rotors used, from left to right. (I - VIII)}

\item{start_positions}{Start positions (e.g. c("A", "A", "A")).}

\item{ring_settings}{Ring settings (e.g. c("D", "E", "F")).}

\item{reflector}{Which reflector is used ("B" or "C").}

\item{plugboard}{Plugboard settings. Should be a named list.}
}
\description{
Available rotors: I-VIII and reflectors B and C.
}
\examples{
# Initialize an enigma machine
enigma <- Enigma(rotors = c("II", "I", "III"), 
                 start_positions = c("D", "E", "F"),
                 ring_settings = c("F", "A", "R"),
                 reflector = "B",
                 plugboard = c(
                     A = "D",
                     C = "Y",
                     F = "Q",
                     G = "R",
                     I = "L",
                     W = "M",
                     Z = "U"
                   )
                )
# Encrypt a few words:
encrypt(enigma, "helloworld")               

}
