% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enigma.R
\docType{class}
\name{Rotor-class}
\alias{Rotor-class}
\alias{Rotor}
\title{An S4 class to represent a rotor in the Enigma machine.}
\description{
An S4 class to represent a rotor in the Enigma machine.
}
\section{Slots}{

\describe{
\item{\code{position}}{Position of the rotor (0-25).}

\item{\code{wiring}}{Wiring of the rotor (relative to the alphabet).}

\item{\code{turnover_position}}{Turnover notch position (A-Z).}

\item{\code{ring_settings}}{Ring setting.}
}}

